dev-lang/ghc: sync with haskell overlay

+++ dev-lang/ghc/ghc-9.8.2.ebuild
@@ -13,7 +13,7 @@ if [[ ${CTARGET} = ${CHOST} ]] ; then
 	fi
 fi
 
-PYTHON_COMPAT=( python3_{10..12} )
+PYTHON_COMPAT=( python3_{9..12} )
 inherit python-any-r1
 inherit autotools bash-completion-r1 flag-o-matic ghc-package
 inherit toolchain-funcs prefix check-reqs llvm unpacker haskell-cabal
@@ -21,7 +21,39 @@ inherit toolchain-funcs prefix check-reqs llvm unpacker haskell-cabal
 DESCRIPTION="The Glasgow Haskell Compiler"
 HOMEPAGE="https://www.haskell.org/ghc/"
 
-SRC_URI="https://downloads.haskell.org/~ghc/${PV}/${P}-src.tar.xz"
+BOOTSTRAP_PV="9.6.2"
+SRC_URI="
+	https://downloads.haskell.org/~ghc/${PV}/${P}-src.tar.xz
+	!ghcbootstrap? (
+		https://downloads.haskell.org/~ghc/9.8.2/hadrian-bootstrap-sources/hadrian-bootstrap-sources-${BOOTSTRAP_PV}.tar.gz
+		amd64? ( https://downloads.haskell.org/~ghc/${BOOTSTRAP_PV}/ghc-${BOOTSTRAP_PV}-x86_64-alpine3_12-linux-static-int_native.tar.xz )
+	)
+"
+
+yet_binary() {
+	case ${ARCH} in
+		amd64)
+			return 0
+			;;
+		*)
+			return 1
+			;;
+	esac
+}
+
+# We are using the upstream static Alpine Linux binaries to bootstrap some
+# archs. These binaries have different properties than the ones we build
+# ourselves, so we need a way to check to see if they are in use.
+upstream_binary() {
+	case ${ARCH} in
+		amd64)
+			return 0
+			;;
+		*)
+			return 1
+			;;
+	esac
+}
 
 GHC_PV=${PV}
 #GHC_PV=8.10.0.20200123 # uncomment only for -alpha, -beta, -rc ebuilds
@@ -35,8 +67,8 @@ BUMP_LIBRARIES=(
 
 LICENSE="BSD"
 SLOT="0/${PV}"
-KEYWORDS="~amd64 ~arm64"
-IUSE="big-endian doc elfutils ghcmakebinary +gmp llvm numa profile test unregisterised"
+#KEYWORDS="~amd64 ~ppc64 ~x86"
+IUSE="big-endian doc elfutils ghcbootstrap ghcmakebinary +gmp llvm numa profile test unregisterised"
 RESTRICT="!test? ( test )"
 
 LLVM_MAX_SLOT="17"
@@ -59,6 +91,7 @@ RDEPEND="
 DEPEND="${RDEPEND}"
 BDEPEND="
 	virtual/pkgconfig
+	dev-util/shake
 	doc? (
 		app-text/docbook-xml-dtd:4.2
 		app-text/docbook-xml-dtd:4.5
@@ -67,8 +100,6 @@ BDEPEND="
 		dev-python/sphinx-rtd-theme
 		>=dev-libs/libxslt-1.1.2
 	)
-	>=dev-lang/ghc-9.4.3
-	dev-util/shake
 	test? ( ${PYTHON_DEPS} )
 "
 
@@ -93,6 +124,10 @@ is_crosscompile() {
 is_native() {
 	[[ ${CHOST} == ${CBUILD} ]] && [[ ${CHOST} == ${CTARGET} ]]
 }
+
+if ! is_crosscompile; then
+	PDEPEND="!ghcbootstrap? ( >=app-admin/haskell-updater-1.2 )"
+fi
 
 # returns tool prefix for crosscompiler.
 # Example:
@@ -223,6 +258,76 @@ ghc_setup_cflags() {
 	done
 }
 
+# substitutes string $1 to $2 in files $3 $4 ...
+relocate_path() {
+	local from=$1
+	local   to=$2
+	shift 2
+	local file=
+	for file in "$@"
+	do
+		sed -i -e "s|$from|$to|g" \
+			"$file" || die "path relocation failed for '$file'"
+	done
+}
+
+# changes hardcoded ghc paths and updates package index
+# $1 - new absolute root path
+relocate_ghc() {
+	local to=$1 ghc_v=${BIN_PV}
+
+	# libdir for prebuilt binary and for current system may mismatch
+	# It does for prefix installation for example: bug #476998
+	local bin_ghc_prefix=${WORKDIR}/usr
+	local bin_libpath=$(echo "${bin_ghc_prefix}"/lib*)
+	local bin_libdir=${bin_libpath#${bin_ghc_prefix}/}
+
+	# backup original script to use it later after relocation
+	local gp_back="${T}/ghc-pkg-${ghc_v}-orig"
+	cp "${WORKDIR}/usr/bin/ghc-pkg-${ghc_v}" "$gp_back" || die "unable to backup ghc-pkg wrapper"
+
+	if [[ ${bin_libdir} != $(get_libdir) ]]; then
+		einfo "Relocating '${bin_libdir}' to '$(get_libdir)' (bug #476998)"
+		# moving the dir itself is not strictly needed
+		# but then USE=binary would result in installing
+		# in '${bin_libdir}'
+		mv "${bin_ghc_prefix}/${bin_libdir}" "${bin_ghc_prefix}/$(get_libdir)" || die
+
+		relocate_path "/usr/${bin_libdir}" "/usr/$(get_libdir)" \
+			"${WORKDIR}/usr/bin/ghc-${ghc_v}" \
+			"${WORKDIR}/usr/bin/ghci-${ghc_v}" \
+			"${WORKDIR}/usr/bin/ghc-pkg-${ghc_v}" \
+			"${WORKDIR}/usr/bin/hsc2hs" \
+			"${WORKDIR}/usr/bin/runghc-${ghc_v}" \
+			"$gp_back" \
+			"${WORKDIR}/usr/$(get_libdir)/${PN}-${ghc_v}/lib/package.conf.d/"*
+	fi
+
+	# Relocate from /usr to ${EPREFIX}/usr
+	relocate_path "/usr" "${to}/usr" \
+		"${WORKDIR}/usr/bin/ghc-${ghc_v}" \
+		"${WORKDIR}/usr/bin/ghci-${ghc_v}" \
+		"${WORKDIR}/usr/bin/ghc-pkg-${ghc_v}" \
+		"${WORKDIR}/usr/bin/haddock-ghc-${ghc_v}" \
+		"${WORKDIR}/usr/bin/hp2ps" \
+		"${WORKDIR}/usr/bin/hpc" \
+		"${WORKDIR}/usr/bin/hsc2hs" \
+		"${WORKDIR}/usr/bin/runghc-${ghc_v}" \
+		"${WORKDIR}/usr/$(get_libdir)/${PN}-${ghc_v}/lib/package.conf.d/"*
+
+	# this one we will use to regenerate cache
+	# so it should point to current tree location
+	relocate_path "/usr" "${WORKDIR}/usr" "$gp_back"
+
+	if use prefix; then
+		hprefixify "${bin_libpath}"/${PN}*/settings
+	fi
+
+	# regenerate the binary package cache
+	"$gp_back" recache || die "failed to update cache after relocation"
+	rm "$gp_back"
+}
+
 ghc-check-reqs() {
 	# These are pessimistic values (slightly bigger than worst-case)
 	# Worst case is UNREG USE=profile ia64. See bug #611866 for some
@@ -328,6 +433,19 @@ src_prepare() {
 
 	ghc_setup_cflags
 
+	if ! use ghcbootstrap && ! upstream_binary; then
+		# Make GHC's settings file comply with user's settings
+	        GHC_SETTINGS="${WORKDIR}/usr/$(get_libdir)/${PN}-${BIN_PV}/lib/settings"
+	        sed -i "s/,(\"C compiler command\", \".*\")/,(\"C compiler command\", \"$(tc-getCC)\")/" "${GHC_SETTINGS}" || die
+	        sed -i "s/,(\"C++ compiler command\", \".*\")/,(\"C++ compiler command\", \"$(tc-getCXX)\")/" "${GHC_SETTINGS}" || die
+	        sed -i "s/,(\"Haskell CPP command\", \".*\")/,(\"Haskell CPP command\", \"$(tc-getCC)\")/" "${GHC_SETTINGS}" || die
+	        sed -i "s/,(\"ld command\", \".*\")/,(\"ld command\", \"$(tc-getLD)\")/" "${GHC_SETTINGS}" || die
+	        sed -i "s/,(\"Merge objects command\", \".*\")/,(\"Merge objects command\", \"$(tc-getLD)\")/" "${GHC_SETTINGS}" || die
+	        sed -i "s/,(\"ar command\", \".*\")/,(\"ar command\", \"$(tc-getAR)\")/" "${GHC_SETTINGS}" || die
+	        sed -i "s/,(\"ranlib command\", \".*\")/,(\"ranlib command\", \"$(tc-getRANLIB)\")/" "${GHC_SETTINGS}" || die
+	fi
+	use llvm && ! use ghcbootstrap && llvmize "${WORKDIR}/usr/bin"
+
 	# binpkg may have been built with FEATURES=splitdebug
 	if [[ -d "${WORKDIR}/usr/lib/debug" ]] ; then
 		rm -rf "${WORKDIR}/usr/lib/debug" || die
@@ -341,17 +459,44 @@ src_prepare() {
 		[[ -e "${f}" ]] || ln -sf "$($(tc-getPKG_CONFIG) --cflags-only-I libffi | sed "s/-I//g" | tr -d " ")/$(basename "${f}")" "${f}" || die
 	done
 
+	if ! use ghcbootstrap && ! upstream_binary; then
+		relocate_ghc "${WORKDIR}"
+	fi
+
 	sed -i -e "s|\"\$topdir\"|\"\$topdir\" ${GHC_PERSISTENT_FLAGS}|" \
 		"${S}/ghc/ghc.wrapper"
 
 	cd "${S}" # otherwise eapply will break
 
+	#eapply "${FILESDIR}"/${PN}-9.0.2-CHOST-prefix.patch
+	#eapply "${FILESDIR}"/${PN}-9.0.2-darwin.patch
+
 	# ModUnusable pretty-printing should include the reason
-	eapply "${FILESDIR}/${PN}-9.0.2-verbose-modunusable.patch"
+	# broken in 9.6.4
+	#eapply "${FILESDIR}/${PN}-9.0.2-verbose-modunusable.patch"
+
+	# Needed for testing with python-3.10
+	#use test && eapply "${FILESDIR}/${PN}-9.0.2-fix-tests-python310.patch"
+
+	#needs a port?
+	#eapply "${FILESDIR}"/${PN}-8.8.1-revert-CPP.patch
+	eapply "${FILESDIR}"/${PN}-8.10.1-allow-cross-bootstrap.patch
+	#eapply "${FILESDIR}"/${PN}-8.10.3-C99-typo-ac270.patch
+	#eapply "${FILESDIR}"/${PN}-9.0.2-disable-unboxed-arrays.patch
+	#eapply "${FILESDIR}"/${PN}-9.0.2-llvm-13.patch
+	#eapply "${FILESDIR}"/${PN}-9.0.2-llvm-14.patch
+
+		# https://gitlab.haskell.org/ghc/ghc/-/issues/22954
+		# https://gitlab.haskell.org/ghc/ghc/-/issues/21936
+		eapply "${FILESDIR}"/${PN}-9.6.4-llvm-16.patch
 
 	# Fix issue caused by non-standard "musleabi" target in
 	# https://gitlab.haskell.org/ghc/ghc/-/blob/ghc-9.4.5-release/m4/ghc_llvm_target.m4#L39
 	eapply "${FILESDIR}"/${PN}-9.4.5-musl-target.patch
+
+	# a bunch of crosscompiler patches
+	# needs newer version:
+	#eapply "${FILESDIR}"/${PN}-8.2.1_rc1-hp2ps-cross.patch
 
 	# build ghc and libraries only the dynamic way
 	eapply "${FILESDIR}"/${PN}-9.4.4-cabal-dynamic-by-default.patch
@@ -396,24 +541,73 @@ src_prepare() {
 }
 
 src_configure() {
-	einfo "Bootstrapping hadrian"
-	( cd "${S}/hadrian" || die
-		local MY_PN="hadrian/hadrian"
-		cabal_chdeps 'Cabal                >= 3.2     && < 3.9' 'Cabal >= 3.2'
-		cabal-bootstrap
-		cabal-configure --flag=-selftest
-		cabal_src_compile || die "Hadrian bootstrap failed"
-	)
+	if ! use ghcbootstrap; then
+		einfo "Installing bootstrap GHC"
+		local ghc_bin_triple
+		case ${ARCH} in
+			amd64)
+				ghc_bin_triple="x86_64-unknown-linux"
+				;;
+			*)
+				die "Unknown ghc binary triple. The list here should match yet_binary."
+				;;
+		esac
+
+		( cd "${WORKDIR}/ghc-${BOOTSTRAP_PV}-${ghc_bin_triple}" || die
+			./configure \
+				--prefix="" \
+				--libdir="/$(get_libdir)" || die
+			emake DESTDIR="${WORKDIR}/ghc-bin" install
+		)
+
+		export PATH="${WORKDIR}/ghc-bin/$(get_libdir)/ghc-${BOOTSTRAP_PV}/bin:${PATH}"
+	fi
+
+        einfo "Bootstrapping hadrian"
+        ( cd "${S}/hadrian" || die
+                local MY_PN="hadrian/hadrian"
+                cabal_chdeps 'Cabal                >= 3.2     && < 3.9' 'Cabal >= 3.2'
+                cabal-bootstrap
+                cabal-configure --flag=-selftest
+                cabal_src_compile || die "Hadrian bootstrap failed"
+        )
 
 	# prepare hadrian build settings files
 	mkdir _build
 	touch _build/hadrian.settings
 
 	# We also need to use the GHC_FLAGS flags when building ghc itself
+	#echo "SRC_HC_OPTS+=${HCFLAGS} ${GHC_FLAGS}" >> mk/build.mk
 	echo "*.*.ghc.hs.opts += ${GHC_FLAGS}" >> _build/hadrian.settings
+	#echo "SRC_CC_OPTS+=${CFLAGS}" >> mk/build.mk
 	# ghc with hadrian is unhappy with these c.opts
 	echo "*.*.ghc.c.opts += ${GHC_FLAGS}" >> _build/hadrian.settings
+	#echo "SRC_LD_OPTS+=${LDFLAGS}" >> mk/build.mk
 #	echo "*.*.ghc.link.opts += ${LDFLAGS}" >> _build/hadrian.settings
+	# Speed up initial Cabal bootstrap
+	#echo "utils/ghc-cabal_dist_EXTRA_HC_OPTS+=$(ghc-make-args)" >> mk/build.mk
+
+#	# not used outside of ghc's test
+#	if [[ -n ${GHC_BUILD_DPH} ]]; then
+#			echo "BUILD_DPH = YES" >> mk/build.mk
+#		else
+#			echo "BUILD_DPH = NO" >> mk/build.mk
+#	fi
+
+#	if is_crosscompile; then
+#		# Install ghc-stage1 crosscompiler instead of
+#		# ghc-stage2 cross-built compiler.
+#		#echo "Stage1Only=YES" >> mk/build.mk
+#		sed -i -e 's/finalStage = Stage2/finalStage = Stage1/' \
+#			hadrian/UserSettings.hs
+#	fi
+
+	# Allow the user to select their bignum backend (default to gmp):
+	# use gmp || sed -i -e 's/userFlavour = defaultFlavour { name = \"user\"/userFlavour = defaultFlavour { name = \"user\", bignumBackend = \"native\"/'
+	#echo "BIGNUM_BACKEND = $(usex gmp gmp native)" >> mk/build.mk
+
+	# don't strip anything. Very useful when stage2 SIGSEGVs on you
+	#echo "STRIP_CMD = :" >> mk/build.mk
 
 	local econf_args=()
 
@@ -435,6 +629,17 @@ src_configure() {
 		# Put docs into the right place, ie /usr/share/doc/ghc-${GHC_PV}
 		--docdir="${EPREFIX}/usr/share/doc/$(cross)${PF}"
 	)
+	case ${CTARGET} in
+		sparc*)
+			# ld.gold-2.28 does not work for ghc. Force ld.bfd
+			# instead. This should be removed once gentoo gets
+			# a fix for missing --no-relax support bug:
+			# https://sourceware.org/ml/binutils/2017-07/msg00183.html
+			econf_args+=(LD=${CTARGET}-ld.bfd)
+		;;
+		*)
+			econf_args+=(LD=${CTARGET}-ld)
+	esac
 
 	if [[ ${CBUILD} != ${CHOST} ]]; then
 		# GHC bug: ghc claims not to support cross-building.
@@ -451,7 +656,9 @@ src_configure() {
 		#  - disable ncurses support for ghci (via haskeline)
 		#    https://bugs.gentoo.org/557478
 		#  - disable ncurses support for ghc-pkg
+		#echo "libraries/haskeline_CONFIGURE_OPTS *. += --flag=-terminfo" >> mk/build.mk
 		echo "*.haskeline.cabal.configure.opts += --flag=-terminfo" >> _build/hadrian.settings
+		#echo "utils/ghc-pkg_HC_OPTS += -DBOOTSTRAPPING" >> mk/build.mk
 		echo "*.ghc-pkg.cabal.configure.opts += --flag=-terminfo" >> _build/hadrian.settings
 	elif is_native; then
 		# using ${GTARGET}'s libffi is not supported yet:
@@ -461,8 +668,10 @@ src_configure() {
 	fi
 
 	einfo "Final _build/hadrian.settings:"
+	#cat mk/build.mk || die
 	cat _build/hadrian.settings || die
 
+#		--enable-bootstrap-with-devel-snapshot \
 	econf ${econf_args[@]} \
 		$(use_enable elfutils dwarf-unwind) \
 		$(use_enable numa) \
@@ -483,18 +692,40 @@ src_compile() {
 	# app-text/dblatex is not in portage, can not build PDF or PS
 	#echo "BUILD_SPHINX_PDF  = NO"  >> mk/build.mk
 	hadrian_vars+=("--docs=no-sphinx-pdfs")
+	#echo "BUILD_SPHINX_HTML = $(usex doc YES NO)" >> mk/build.mk
 	use doc || hadrian_vars+=("--docs=no-sphinx-html")
+	#echo "BUILD_MAN = $(usex doc YES NO)" >> mk/build.mk
 	use doc || hadrian_vars+=("--docs=no-sphinx-man")
+	# this controls presence on 'xhtml' and 'haddock' in final install
+	#echo "HADDOCK_DOCS       = YES" >> mk/build.mk
 	use doc || hadrian_vars+=("--docs=no-haddocks")
 
 	# Any non-native build has to skip as it needs
 	# target haddock binary to be runnabine.
 	if ! is_native; then
 		# disable docs generation as it requires running stage2
+		# echo "HADDOCK_DOCS=NO" >> mk/build.mk
 		hadrian_vars+=("--docs=no-haddocks")
+		# echo "BUILD_SPHINX_HTML=NO" >> mk/build.mk
 		hadrian_vars+=("--docs=no-sphinx-pdfs")
+		# echo "BUILD_SPHINX_PDF=NO" >> mk/build.mk
 		hadrian_vars+=("--docs=no-sphinx-html")
 	fi
+
+#	# allows overriding build flavours for libraries:
+#	# v   - vanilla (static libs)
+#	# p   - profiled
+#	# dyn - shared libraries
+#	# example: GHC_LIBRARY_WAYS="v dyn"
+#	if [[ -n ${GHC_LIBRARY_WAYS} ]]; then
+#		echo "GhcLibWays=${GHC_LIBRARY_WAYS}" >> mk/build.mk
+#	fi
+#	echo "BUILD_PROF_LIBS = $(usex profile YES NO)" >> mk/build.mk
+
+	###
+	# TODO: Move these env vars to a hadrian eclass, for better
+	# documentation and clarity
+	###
 
 	# Control the build flavour
 	if use profile; then
@@ -505,7 +736,7 @@ src_compile() {
 
 	hadrian_vars+=("--flavour=${HADRIAN_FLAVOUR}")
 
-	# Control the verbosity of hadrian. Default is one level of --verbose
+	# Control the verbosity of hadrian. Default is two levels of --verbose
 	${HADRIAN_VERBOSITY:=2}
 
 	local n="${HADRIAN_VERBOSITY}"
@@ -521,6 +752,24 @@ src_compile() {
 		esac
 	done
 
+#	# Stage1Only crosscompiler does not build stage2
+#	if ! is_crosscompile; then
+#		# 1. build/pax-mark compiler binary first
+#		#emake ghc/stage2/build/tmp/ghc-stage2
+#		hadrian -j${nproc} --flavour=quickest stage2:exe:ghc-bin || die
+#		# 2. pax-mark (bug #516430)
+#		#pax-mark -m _build/stage1/bin/ghc
+#		# 2. build/pax-mark haddock using ghc-stage2
+#		if is_native; then
+#			# non-native build does not build haddock
+#			# due to HADDOCK_DOCS=NO, but it could.
+#			#emake utils/haddock/dist/build/tmp/haddock
+#			hadrian docs --docs=no-sphinx-pdfs --docs=no-sphinx-html || die
+#			#pax-mark -m utils/haddock/dist/build/tmp/haddock
+#		fi
+#	fi
+#	# 3. and then all the rest
+#	#emake all
 
 	local hadrian=( "${S}/hadrian/dist/build/hadrian/hadrian" )
 	hadrian+=(
@@ -537,6 +786,9 @@ src_test() {
 	#    - sandbox (pollutes environment)
 	#    - extra packages (to extend testsuite coverage)
 	# bits are taken from 'validate'
+	#local make_test_target='test' # can be fulltest
+	# not 'emake' as testsuite uses '$MAKE' without jobserver available
+	#make $make_test_target stage=2 THREADS=$(makeopts_jobs)
 	hadrian test || die
 }
 
@@ -545,10 +797,15 @@ src_install() {
 
 	[[ -f VERSION ]] || emake VERSION
 
+#	einfo "Running: hadrian install ${hadrian_vars}"
+#	hadrian install --prefix="${D}/usr/" ${hadrian_vars} || die
+
 	pushd "${S}/_build/bindist/${P}-${CHOST}" || die
 	econf
 	emake DESTDIR="${D}" install
 	popd
+
+	#emake -j1 install DESTDIR="${D}"
 
 	use llvm && llvmize "${ED}/usr/bin"
